db: add migrations
